+++ util/blobs/blobs.go
@@ -1,0 +1,28 @@
+package blobs
+
+import (
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/params"
+)
+
+// EncodeBlobs takes in raw bytes data to convert into blobs used for KZG commitment EIP-4844
+// transactions on Ethereum.
+func EncodeBlobs(data []byte) types.Blobs {
+	blobs := []types.Blob{{}}
+	blobIndex := 0
+	fieldIndex := -1
+	for i := 0; i < len(data); i += 31 {
+		fieldIndex++
+		if fieldIndex == params.FieldElementsPerBlob {
+			blobs = append(blobs, types.Blob{})
+			blobIndex++
+			fieldIndex = 0
+		}
+		max := i + 31
+		if max > len(data) {
+			max = len(data)
+		}
+		copy(blobs[blobIndex][fieldIndex][:], data[i:max])
+	}
+	return blobs
+}
